Fix Dask viz pickle bug by routing visualization through the task manager
Post-run visualization in BaseSimulator parallelized via joblib.Parallel
with a closure that captured `self`, transitively pulling in the live
dask.distributed.Client and its `_asyncio.Task` — unpicklable, so Dask
runs aborted with `TypeError: cannot pickle '_asyncio.Task' object`
after the simulation phase had already succeeded.

This commit:

* Introduces an ExperimentVisualizer ABC under core/simulation/ and a
  concrete EpisodeReturnsVisualizer holding the per-policy histogram,
  comparison-histogram, and per-episode env-cache logic that previously
  lived as private methods on POMDPSimulator.

* Adds an EnvironmentVisualizationTask SimulationTask subclass; the
  simulator dispatches one per environment via self.task_manager.run_tasks,
  using the same backend (Dask / Joblib / PBS / Sequential) as the
  simulation phase. Visualization now scales across the full cluster on
  Dask runs instead of being locked to one machine's joblib pool.

* Splits simulator.py (1,218 lines) into a simulator/ package
  (base_simulator.py, pomdp_simulator.py, episode_returns_visualizer.py)
  with __init__.py re-exports so all existing import paths keep working.

* Migrates the broken viz tests in test_simulator.py, adds dedicated
  unit tests for the new abstractions, and adds an end-to-end Dask
  regression test that runs the full pipeline against a LocalCluster
  and asserts no pickle errors.
