+++ POMDPPlanners/simulations/simulator/__init__.py
@@ -1,0 +1,14 @@
+"""POMDP simulator package.
+
+Re-exports the public simulator API. Importers can continue to use::
+
+    from POMDPPlanners.simulations.simulator import POMDPSimulator
+"""
+
+from POMDPPlanners.simulations.simulator.base_simulator import BaseSimulator
+from POMDPPlanners.simulations.simulator.episode_returns_visualizer import (
+    EpisodeReturnsVisualizer,
+)
+from POMDPPlanners.simulations.simulator.pomdp_simulator import POMDPSimulator
+
+__all__ = ["BaseSimulator", "POMDPSimulator", "EpisodeReturnsVisualizer"]
